ports/UART: Use 32 bit values for timeout and timeout_char.
The previous 16 bit values limit the timeout to ~65 seconds, which
seemed sufficiently long so far, especially since it is a blocking
wait.
Implemented and tested for the ports STM32, ESP32, ESP8266, RP2,
MIMXRT, SAMD, Renesas-RA and nrf.

+++ ports/esp32/machine_uart.c
@@ -65,8 +65,8 @@ typedef struct _machine_uart_obj_t {
     gpio_num_t cts;
     uint16_t txbuf;
     uint16_t rxbuf;
-    uint16_t timeout;       // timeout waiting for first char (in ms)
-    uint16_t timeout_char;  // timeout waiting between chars (in ms)
+    uint32_t timeout;       // timeout waiting for first char (in ms)
+    uint32_t timeout_char;  // timeout waiting between chars (in ms)
     uint32_t invert;        // lines to invert
 } machine_uart_obj_t;
